照共功能分試驗
