+++ 試驗/解析整理/Test拆文分析器連字符單元試驗.py
@@ -4,32 +4,13 @@ from 臺灣言語工具.解析整理.拆文分析器 import 拆文分析器
 from 臺灣言語工具.基本物件.字 import 字
 
 
-class 拆文分析器對齊連字符單元試驗(TestCase):
+class 拆文分析器連字符單元試驗(TestCase):
     def test_句頭輕聲無連做伙嘛會使(self):
         型 = '-- 啊'
         音 = '-- ah'
         組物件 = 拆文分析器.對齊組物件(型, 音)
         self.assertEqual(組物件.篩出字物件(), [
             字('啊', '0ah'),
-        ])
-
-    def test_句中輕聲無連做伙嘛會使(self):
-        漢 = '講會出 -- 來'
-        羅 = 'āu--ji̍t'
-        組物件 = 拆文分析器.對齊組物件(漢, 羅)
-        self.assertEqual(len(組物件.網出詞物件()), 1)
-        self.assertEqual(組物件.篩出字物件(), [
-            字('後', 'āu'),
-            字('日', '0ji̍t'),
-        ])
-
-    def test_句中輕聲kah4後壁無連做伙嘛會使(self):
-        漢 = '講 -- 出 - 來'
-        羅 = 'āu--ji̍t'
-        組物件 = 拆文分析器.對齊組物件(漢, 羅)
-        self.assertEqual(len(組物件.網出詞物件()), 2)
-        self.assertEqual(組物件.網出詞物件()[1].篩出字物件(), [
-            字('矣', '0ah'),
         ])
 
     def test_有處理減號的輕聲(self):
